ResetUserPassword cli migration
